+++ src/commands/openDocumentLink.ts
@@ -8,7 +8,7 @@ import { extname } from 'path';
 
 import { Command } from '../commandManager';
 import { TextileEngine } from '../textileEngine';
-//import { TableOfContentsProvider } from '../tableOfContentsProvider';
+import { TableOfContentsProvider } from '../tableOfContentsProvider';
 import { isTextileFile } from '../util/file';
 
 
@@ -85,13 +85,11 @@ export class OpenDocumentLinkCommand implements Command {
 
 	private async tryRevealLine(editor: vscode.TextEditor, fragment?: string) {
 		if (editor && fragment) {
-			/* FIXME : supporter ?
 			const toc = new TableOfContentsProvider(this.engine, editor.document);
 			const entry = await toc.lookup(fragment);
 			if (entry) {
 				return editor.revealRange(new vscode.Range(entry.line, 0, entry.line, 0), vscode.TextEditorRevealType.AtTop);
 			}
-			*/
 			const lineNumberFragment = fragment.match(/^L(\d+)$/i);
 			if (lineNumberFragment) {
 				const line = +lineNumberFragment[1] - 1;