+++ src/extension.ts
@@ -11,17 +11,17 @@ nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFo
 import { CommandManager } from './commandManager';
 import * as commands from './commands/index';
 import LinkProvider from './features/documentLinkProvider';
-//import TextileDocumentSymbolProvider from './features/documentSymbolProvider';
-//import TextileFoldingProvider from './features/foldingProvider';
+import TextileDocumentSymbolProvider from './features/documentSymbolProvider';
+import TextileFoldingProvider from './features/foldingProvider';
 import { TextileContentProvider } from './features/previewContentProvider';
 import { TextilePreviewManager } from './features/previewManager';
-//import TextileWorkspaceSymbolProvider from './features/workspaceSymbolProvider';
+import TextileWorkspaceSymbolProvider from './features/workspaceSymbolProvider';
 import { Logger } from './logger';
 import { TextileEngine } from './textileEngine';
 import { getTextileExtensionContributions } from './textileExtensions';
 import { ExtensionContentSecurityPolicyArbiter, PreviewSecuritySelector, ContentSecurityPolicyArbiter } from './security';
 //import { loadDefaultTelemetryReporter, TelemetryReporter } from './telemetryReporter';
-//import { githubSlugifier } from './slugify';
+import { githubSlugifier } from './slugify';
 
 
 export function activate(context: vscode.ExtensionContext) {
@@ -34,26 +34,15 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(contributions);
 
 	const cspArbiter = new ExtensionContentSecurityPolicyArbiter(context.globalState, context.workspaceState);
-	const engine = new TextileEngine(contributions
-		/* FIXME activate ? Need changes inside textile-js :
-			, githubSlugifier
-		*/
-	);
+	const engine = new TextileEngine(contributions, githubSlugifier);
 	const logger = new Logger();
 
 	const contentProvider = new TextileContentProvider(engine, context, cspArbiter, contributions, logger);
-	/* FIXME : activate
 	const symbolProvider = new TextileDocumentSymbolProvider(engine);
-	*/
 	const previewManager = new TextilePreviewManager(contentProvider, logger, contributions);
 	context.subscriptions.push(previewManager);
 
-	context.subscriptions.push(registerTextileLanguageFeatures(
-		/* FIXME : activate
-		symbolProvider,
-		*/
-		engine
-	));
+	context.subscriptions.push(registerTextileLanguageFeatures(symbolProvider, engine));
 	context.subscriptions.push(registerTextileCommands(previewManager, /* Disabled for textile : telemetryReporter, */ cspArbiter, engine));
 
 	context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(() => {
@@ -63,9 +52,7 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 function registerTextileLanguageFeatures(
-	/* FIXME : activate
 	symbolProvider: TextileDocumentSymbolProvider,
-	*/
 	engine: TextileEngine
 ): vscode.Disposable {
 	const selector: vscode.DocumentSelector = { language: 'textile', scheme: '*' };
@@ -76,14 +63,10 @@ function registerTextileLanguageFeatures(
 		vscode.languages.setLanguageConfiguration('textile', {
 			wordPattern: new RegExp(`${charPattern}((${charPattern}|[_])?${charPattern})*`, 'ug'),
 		}),
-		/* FIXME : activate
 		vscode.languages.registerDocumentSymbolProvider(selector, symbolProvider),
-		*/
-		vscode.languages.registerDocumentLinkProvider(selector, new LinkProvider())
-		/* FIXME : activate
+		vscode.languages.registerDocumentLinkProvider(selector, new LinkProvider()),
 		vscode.languages.registerFoldingRangeProvider(selector, new TextileFoldingProvider(engine)),
 		vscode.languages.registerWorkspaceSymbolProvider(new TextileWorkspaceSymbolProvider(symbolProvider))
-		*/
 	);
 }