+++ src/textileEngine.ts
@@ -5,23 +5,21 @@
 
 //import * as crypto from 'crypto';
 //import * as path from 'path';
-import { TextileJS, Token } from '../libs/textile-js/textile';
+import { TextileJS, Token, Options as TextileJSConfig } from '../libs/textile-js/textile';
 import * as vscode from 'vscode';
 import { TextileContributionProvider as TextileContributionProvider } from './textileExtensions';
-//import { Slugifier } from './slugify';
+import { Slugifier } from './slugify';
 import { SkinnyTextDocument } from './tableOfContentsProvider';
 //import { Schemes, isOfScheme } from './util/links';
 
 const UNICODE_NEWLINE_REGEX = /\u2028|\u2029/g;
 
-// -- Begin : Changed for textile
+/* Disabled for textile : already defined in textile lib
 interface TextileJSConfig {
 	readonly breaks: boolean;
-	/* Disabled for Textile : readonly linkify: boolean; */
-	readonly showOriginalLineNumber: boolean;
-	readonly cssClassOriginalLineNumber: string;
+	readonly linkify: boolean;
 }
-// -- End : Changed for textile
+*/
 
 class TokenCache {
 	private cachedDocument?: {
@@ -64,12 +62,12 @@ export class TextileEngine {
 	private textile?: Promise<TextileJS>;
 
 	// Disabled for textile : private currentDocument?: vscode.Uri;
-	// FIXME : private _slugCount = new Map<string, number>();
+	private _slugCount = new Map<string, number>();
 	private _tokenCache = new TokenCache();
 
 	public constructor(
 		private readonly contributionProvider: TextileContributionProvider,
-		/* FIXME activate : private readonly slugifier: Slugifier,*/
+		private readonly slugifier: Slugifier,
 	) {
 		contributionProvider.onContributionsChanged(() => {
 			// Textile plugin contributions may have changed
@@ -79,9 +77,8 @@ export class TextileEngine {
 
 	private async getEngine(config: TextileJSConfig): Promise<TextileJS> {
 		if (!this.textile) {
-			this.textile = import('../libs/textile-js/textile');
-			/* Disabled for Textile :
-			.then(async textileIt => {
+			this.textile = import('../libs/textile-js/textile').then(async textile => {
+				/* -- Begin : changed for Textile :
 				let md: TextileJS = textileIt(await getTextileOptions(() => md));
 
 				for (const plugin of this.contributionProvider.contributions.textileItPlugins.values()) {
@@ -110,16 +107,23 @@ export class TextileEngine {
 				for (const renderName of ['paragraph_open', 'heading_open', 'image', 'code_block', 'fence', 'blockquote_open', 'list_item_open']) {
 					this.addLineNumberRenderer(md, renderName);
 				}
+				*/
 
-				this.addImageStabilizer(md);
-				this.addFencedRenderer(md);
-				this.addLinkNormalizer(md);
-				this.addLinkValidator(md);
-				this.addNamedHeaders(md);
-				this.addLinkRenderer(md);
-				return md;
+				// hooks are set only once : don't add them to config member parameter
+				const localConfig :TextileJSConfig = {
+					hooks: []
+				};
+				// FIXME ? this.addImageStabilizer(md);
+				// disabled for textile : this.addFencedRenderer(md);
+				// FIXME ? this.addLinkNormalizer(md);
+				// FIXME ? this.addLinkValidator(md);
+				this.addNamedHeaders(textile, localConfig);
+				// disabled for textile : this.addLinkRenderer(md);
+				textile.setOptions( localConfig );
+				// -- End : changed for textile
+
+				return textile;
 			});
-			*/
 		}
 
 		const textile = await this.textile!;
@@ -140,6 +144,24 @@ export class TextileEngine {
 	}
 	// -- End: Keep for Textile
 
+	// -- Begin: Added for Textile
+	public async jsonmlUtils() :Promise<TextileJS["jsonmlUtils"]> {
+		return (await this.textile!).jsonmlUtils;
+	}
+
+	private reduceTokenString(jsonml: Token[]) :string {
+		let res = '';
+		for( let i = 1, l = jsonml.length; i < l; i++) {
+			if (typeof( jsonml[i] ) === 'string') {
+				res += jsonml[i];
+			} else if (Array.isArray(jsonml[i])) {
+				res += this.reduceTokenString(jsonml[i]);
+			}
+		}
+		return res;
+	}
+	// -- End: Added for Textile
+
 	private tokenizeDocument(
 		document: SkinnyTextDocument,
 		config: TextileJSConfig,
@@ -151,7 +173,7 @@ export class TextileEngine {
 		}
 
 		// Disabled for textile : this.currentDocument = document.uri;
-		// FIXME : this._slugCount = new Map<string, number>();
+		this._slugCount = new Map<string, number>();
 
 		const tokens = this.tokenizeString(document.getText(), engine);
 		this._tokenCache.update(document, config, tokens);
@@ -204,7 +226,7 @@ export class TextileEngine {
 		// -- End : Changed for textile
 	}
 
-	/* FIXME : not used for textile
+	/* FIXME
 	private addLineNumberRenderer(md: any, ruleName: string): void {
 		const original = md.renderer.rules[ruleName];
 		md.renderer.rules[ruleName] = (tokens: any, idx: number, options: any, env: any, self: any) => {
@@ -309,32 +331,41 @@ export class TextileEngine {
 				|| /^data:image\/.*?;/.test(link);
 		};
 	}
+*/
 
-	private addNamedHeaders(md: any): void {
-		const original = md.renderer.rules.heading_open;
-		md.renderer.rules.heading_open = (tokens: any, idx: number, options: any, env: any, self: any) => {
-			const title = tokens[idx + 1].children.reduce((acc: string, t: any) => acc + t.content, '');
-			let slug = this.slugifier.fromHeading(title);
+	// -- Begin : Changed for textile
+	private addNamedHeaders(textile: TextileJS, config: TextileJSConfig): void {
+		config.hooks!.push(
+			 [(tokens: Token[]) => {
+				switch( tokens[0] ) {
+					case 'h1':
+					case 'h2':
+					case 'h3':
+					case 'h4':
+					case 'h5':
+					case 'h6':
+						const title = this.reduceTokenString( tokens );
+						let slug = this.slugifier.fromHeading(title);
 
-			if (this._slugCount.has(slug.value)) {
-				const count = this._slugCount.get(slug.value)!;
-				this._slugCount.set(slug.value, count + 1);
-				slug = this.slugifier.fromHeading(slug.value + '-' + (count + 1));
-			} else {
-				this._slugCount.set(slug.value, 0);
-			}
-
-			tokens[idx].attrs = tokens[idx].attrs || [];
-			tokens[idx].attrs.push(['id', slug.value]);
-
-			if (original) {
-				return original(tokens, idx, options, env, self);
-			} else {
-				return self.renderToken(tokens, idx, options, env, self);
-			}
-		};
+						if (this._slugCount.has(slug.value)) {
+							const count = this._slugCount.get(slug.value)!;
+							this._slugCount.set(slug.value, count + 1);
+							slug = this.slugifier.fromHeading(slug.value + '-' + (count + 1));
+						} else {
+							this._slugCount.set(slug.value, 0);
+						}
+						textile.jsonmlUtils.addAttributes( tokens, { 'id': slug.value } );
+						break;
+					default:
+						break;
+				}
+				return tokens;
+			}]
+		);
 	}
+	// -- End : Changed for textile
 
+	/* FIXME ?
 	private addLinkRenderer(md: any): void {
 		const old_render = md.renderer.rules.link_open || ((tokens: any, idx: number, options: any, _env: any, self: any) => {
 			return self.renderToken(tokens, idx, options);